Use acme's jose where present, then josepy
Within jose, there is an assertion when signing the challenge that the
key is an instance of the right sort of key. This assertion fails if
autocert and acme are using different jose packages.

I think this solution should allow autocert to be compatible with more
recent versions of acme (which use josepy) and the older versions (which
include a jose package)